refactor router helper and update prevUrl&nextUrl method
